move to test_server.py
