+++ tests/test_config.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import asyncio
 import configparser
 import io
 import json
@@ -13,12 +12,11 @@ from pathlib import Path
 from typing import Any, Literal
 from unittest.mock import MagicMock
 
-import httpx
 import pytest
 import yaml
 from pytest_mock import MockerFixture
 
-from tests.utils import as_cwd, run_server
+from tests.utils import as_cwd
 from uvicorn._types import (
     ASGIApplication,
     ASGIReceiveCallable,
@@ -31,9 +29,6 @@ from uvicorn.config import Config
 from uvicorn.middleware.proxy_headers import ProxyHeadersMiddleware
 from uvicorn.middleware.wsgi import WSGIMiddleware
 from uvicorn.protocols.http.h11_impl import H11Protocol
-from uvicorn.protocols.http.httptools_impl import HttpToolsProtocol
-
-pytestmark = pytest.mark.anyio
 
 
 @pytest.fixture
@@ -57,12 +52,6 @@ async def asgi_app(scope: Scope, receive: ASGIReceiveCallable, send: ASGISendCal
 
 def wsgi_app(environ: Environ, start_response: StartResponse) -> None:
     pass  # pragma: nocover
-
-
-async def app(scope: Scope, receive: ASGIReceiveCallable, send: ASGISendCallable) -> None:
-    assert scope["type"] == "http"
-    await send({"type": "http.response.start", "status": 200, "headers": []})
-    await send({"type": "http.response.body", "body": b"", "more_body": False})
 
 
 @pytest.mark.parametrize(
@@ -556,16 +545,3 @@ def test_warn_when_using_reload_and_workers(caplog: pytest.LogCaptureFixture) ->
     Config(app=asgi_app, reload=True, workers=2)
     assert len(caplog.records) == 1
     assert '"workers" flag is ignored when reloading is enabled.' in caplog.records[0].message
-
-
-async def test_request_than_limit_max_requests_warn_log(
-    unused_tcp_port: int, http_protocol_cls: type[H11Protocol | HttpToolsProtocol], caplog: pytest.LogCaptureFixture
-):
-    caplog.set_level(logging.WARNING, logger="uvicorn.error")
-    config = Config(app=app, limit_max_requests=1, port=unused_tcp_port, http=http_protocol_cls)
-    async with run_server(config):
-        async with httpx.AsyncClient() as client:
-            tasks = [client.get(f"http://127.0.0.1:{unused_tcp_port}") for _ in range(2)]
-            responses = await asyncio.gather(*tasks)
-            assert len(responses) == 2
-    assert f"Maximum request limit of {config.limit_max_requests} exceeded. Terminating process." in caplog.text
